+++ components/version-info.tsx
